Crate sc_utils

Utilities Primitives for Substrate

Features

metered

This feature changes the behaviour of the function mpsc::tracing_unbounded. With the disabled feature this function is an alias to futures::channel::mpsc::unbounded. However, when the feature is enabled it creates wrapper types to UnboundedSender<T> and UnboundedReceiver<T> to register every send/received/dropped action happened on the channel.

Also this feature creates and registers a prometheus vector with name unbounded_channel_len and labels:

Label	Description

entity	Name of channel passed to tracing_unbounded

action	One of send/received/dropped

Modules

id_sequence

Produce opaque sequential IDs.

metrics

Metering primitives and globals

mpsc

Features to meter unbounded channels

notification

Provides mpsc notification channel that can be instantiated after it’s been shared to the consumer and producers entities.

pubsub

Provides means to implement a typical Pub/Sub mechanism.

status_sinks