sc_utils

Module mpsc

Code to meter unbounded channels.

Structs

TracingUnboundedReceiver

Wrapper Type around async_channel::Receiver that decreases the global measure when a message is polled.

TracingUnboundedSender

Wrapper Type around async_channel::Sender that increases the global measure when a message is added.

Enums

TryRecvError

An error returned from Receiver::try_recv().

TrySendError

An error returned from Sender::try_send().

Functions

tracing_unbounded

Wrapper around async_channel::unbounded that tracks the in- and outflow via UNBOUNDED_CHANNELS_COUNTER and warns if the message queue grows above the warning threshold.