Trait sc_utils::notification::TracingKeyStr

pub trait TracingKeyStr {
    const TRACING_KEY: &'static str;
}

Trait used to define the “tracing key” string used to tag and identify the mpsc channels.

Associated Constants

const TRACING_KEY: &'static str

Const str representing the “tracing key” used to tag and identify the mpsc channels owned by the object implemeting this trait.