Struct sc_utils::pubsub::Hub

pub struct Hub<M, R> { /* private fields */ }

A subscription hub.

Does the subscription and dispatch. The exact subscription and routing behaviour is to be implemented by the Registry (of type R). The Hub under the hood uses the channel defined in crate::mpsc module.

Implementations

impl<M, R> Hub<M, R>where R: Unsubscribe,

pub fn map_registry_for_tests<MapF, Ret>(&self, map: MapF) -> Retwhere MapF: FnOnce(&R) -> Ret,

Provide access to the registry (for test purposes).

impl<M, R> Hub<M, R>

pub fn new(tracing_key: &'static str) -> Selfwhere R: Default,

Create a new instance of Hub (with default value for the Registry).

pub fn new_with_registry(tracing_key: &'static str, registry: R) -> Self

Create a new instance of Hub over the initialized Registry.

pub fn subscribe<K>( &self, subs_key: K, queue_size_warning: usize ) -> Receiver<M, R>where R: Subscribe<K> + Unsubscribe,

Subscribe to this Hub using the subs_key: K.

A subscription with a key K is possible if the Registry implements Subscribe<K>.

pub fn send<Trigger>(&self, trigger: Trigger) -> <R as Dispatch<Trigger>>::Retwhere R: Dispatch<Trigger, Item = M>,

Send the message produced with Trigger.

This is possible if the registry implements Dispatch<Trigger, Item = M>.