Module sc_utils::mpsc

Features to meter unbounded channels

Structs

TracingUnboundedReceiver

Wrapper Type around UnboundedReceiver that decreases the global measure when a message is polled

TracingUnboundedSender

Wrapper Type around UnboundedSender that increases the global measure when a message is added

Functions

tracing_unbounded

Wrapper around mpsc::unbounded that tracks the in- and outflow via UNBOUNDED_CHANNELS_COUNTER