Module sc_utils::notification

Provides mpsc notification channel that can be instantiated after it’s been shared to the consumer and producers entities.

Useful when building RPC extensions where, at service definition time, we don’t know whether the specific interface where the RPC extension will be exposed is safe or not and we want to lazily build the RPC extension whenever we bind the service to an interface.

See [sc-service::builder::RpcExtensionBuilder] for more details.

Structs

NotificationSender

The sending half of the notifications channel(s).

NotificationStream

The receiving half of the notifications channel.

Traits

TracingKeyStr

Trait used to define the “tracing key” string used to tag and identify the mpsc channels.